refactor: single node connector
