+++ packages/mosaic/core/package.json
@@ -13,8 +13,14 @@
   "author": "Jeffrey Heer (https://idl.uw.edu)",
   "type": "module",
   "exports": {
-    "types": "./dist/src/index.d.ts",
-    "default": "./dist/src/index.js"
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "default": "./dist/src/index.js"
+    },
+    "./node-connector": {
+      "types": "./dist/src/connectors/NodeConnector.d.ts",
+      "default": "./dist/src/connectors/NodeConnector.js"
+    }
   },
   "files": [
     "dist",
@@ -39,5 +45,13 @@
   },
   "devDependencies": {
     "@uwdata/mosaic-duckdb": "workspace:^"
+  },
+  "peerDependencies": {
+    "@uwdata/mosaic-duckdb": "workspace:^"
+  },
+  "peerDependenciesMeta": {
+    "@uwdata/mosaic-duckdb": {
+      "optional": true
+    }
   }
 }
