+++ packages/mosaic/core/src/connectors/NodeConnector.ts
@@ -1,10 +1,17 @@
-
 import type { ExtractionOptions, Table } from '@uwdata/flechette';
 import { DuckDB } from '@uwdata/mosaic-duckdb';
-import type { ArrowQueryRequest, Connector, ExecQueryRequest, JSONQueryRequest } from '../../src/index.js';
-import { decodeIPC } from '../../src/util/decode-ipc.js';
-import { ConnectorQueryRequest } from '../../src/connectors/Connector.js';
+import { decodeIPC } from '../util/decode-ipc.js';
+import type {
+  ArrowQueryRequest,
+  Connector,
+  ConnectorQueryRequest,
+  ExecQueryRequest,
+  JSONQueryRequest
+} from './Connector.js';
 
+/**
+ * A Mosaic Connector backed by an in-process Node.js DuckDB instance.
+ */
 export class NodeConnector implements Connector {
   protected _db: DuckDB;
   protected _ipc?: ExtractionOptions;
