+++ packages/mosaic/core/test/client.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import type { Table } from '@uwdata/flechette';
 import { Query, count } from '@uwdata/mosaic-sql';
-import { NodeConnector } from './util/node-connector.js';
+import { NodeConnector } from '../src/connectors/NodeConnector.js';
 import { Coordinator, MosaicClient, Selection, clauseInterval } from '../src/index.js';
 import { QueryResult } from '../src/util/query-result.js';
